+++ spec/domain/export/pdf/messages/letter_with_invoice_spec.rb
@@ -8,6 +8,8 @@
 require "spec_helper"
 
 describe Export::Pdf::Messages::LetterWithInvoice do
+  include PdfHelpers
+
   let(:letter) { messages(:with_invoice) }
   let(:options) { {} }
   let(:group) { groups(:top_layer) }
@@ -51,8 +53,7 @@ describe Export::Pdf::Messages::LetterWithInvoice do
     subject { PDF::Inspector::Text.analyze(pdf.render) }
 
     it "renders text at positions" do
-      described_class.new(letter).render
-      expect(text_with_position).to match_array [
+      invoice_text = [
         [71, 654, "Bottom Member"],
         [71, 644, "Greatstreet 345"],
         [71, 633, "3456 Greattown"],
@@ -64,43 +65,48 @@ describe Export::Pdf::Messages::LetterWithInvoice do
         [71, 460, "Bis bald"],
         [28, 290, "Empfangsschein"],
         [28, 265, "Konto / Zahlbar an"],
-        [28, 254, "CH93 0076 2011 6238 5295 7"],
-        [28, 242, "Hans Gerber"],
-        [28, 230, "Eine Strasse 42"],
-        [28, 219, "1234 Dorf"],
-        [28, 187, "Zahlbar durch"],
-        [28, 175, "Bottom Member"],
-        [28, 164, "Greatstreet 345"],
+        [28, 255, "CH93 0076 2011 6238 5295 7"],
+        [28, 246, "Hans Gerber"],
+        [28, 237, "Eine Strasse 42"],
+        [28, 228, "1234 Dorf"],
+        [28, 208, "Referenznummer"],
+        [28, 199, "00 00834 96356 70000 00000 00019"],
+        [28, 180, "Zahlbar durch"],
+        [28, 170, "Bottom Member"],
+        [28, 161, "Greatstreet 345"],
         [28, 152, "3456 Greattown"],
         [28, 103, "Währung"],
         [85, 103, "Betrag"],
-        [28, 92, "CHF"],
-        [85, 92, "10.00"],
+        [28, 91, "CHF"],
+        [85, 91, "10.00"],
         [119, 53, "Annahmestelle"],
         [204, 290, "Zahlteil"],
-        [204, 103, "Währung"],
-        [261, 103, "Betrag"],
-        [204, 92, "CHF"],
-        [261, 92, "10.00"],
-        [360, 292, "Konto / Zahlbar an"],
-        [360, 280, "CH93 0076 2011 6238 5295 7"],
-        [360, 269, "Hans Gerber"],
-        [360, 257, "Eine Strasse 42"],
-        [360, 246, "1234 Dorf"],
-        [360, 225, "Referenznummer"],
-        [360, 214, "00 00834 96356 70000 00000 00019"],
-        [360, 193, "Zahlbar durch"],
-        [360, 181, "Bottom Member"],
-        [360, 170, "Greatstreet 345"],
-        [360, 158, "3456 Greattown"]
+        [204, 102, "Währung"],
+        [261, 102, "Betrag"],
+        [204, 90, "CHF"],
+        [261, 90, "10.00"],
+        [360, 290, "Konto / Zahlbar an"],
+        [360, 279, "CH93 0076 2011 6238 5295 7"],
+        [360, 267, "Hans Gerber"],
+        [360, 255, "Eine Strasse 42"],
+        [360, 244, "1234 Dorf"],
+        [360, 222, "Referenznummer"],
+        [360, 210, "00 00834 96356 70000 00000 00019"],
+        [360, 189, "Zahlbar durch"],
+        [360, 177, "Bottom Member"],
+        [360, 165, "Greatstreet 345"],
+        [360, 154, "3456 Greattown"]
       ]
+
+      invoice_text.each_with_index do |text, i|
+        expect(text_with_position[i]).to eq(text)
+      end
     end
 
     context "persisted invoice list" do
       it "renders iban from invoice config when no persisted invoice exists" do
         group.invoice_config.update(iban: "CH10 0221 1981 6169 5329 8")
-        described_class.new(letter).render
-        expect(text_with_position).to include([360, 280, "CH10 0221 1981 6169 5329 8"])
+        expect(text_with_position).to include([360, 279, "CH10 0221 1981 6169 5329 8"])
       end
 
       it "renders iban from invoice when persisted invoice exists" do
@@ -109,8 +115,7 @@ describe Export::Pdf::Messages::LetterWithInvoice do
         letter.invoice_list_id = list.id
 
         group.invoice_config.update!(iban: "CH10 0221 1981 6169 5329 8")
-        described_class.new(letter).render
-        expect(text_with_position).to include([360, 280, "CH93 0076 2011 6238 5295 7"])
+        expect(text_with_position).to include([360, 279, "CH93 0076 2011 6238 5295 7"])
       end
 
       it "mixes person and invoice address" do
@@ -122,8 +127,7 @@ describe Export::Pdf::Messages::LetterWithInvoice do
 
         described_class.new(letter).render
         expect(text_with_position).to include([71, 654, "Foo Member"])
-        expect(text_with_position).to include([28, 175, "Bottom Member"])
-        expect(text_with_position).to include([360, 181, "Bottom Member"])
+        expect(text_with_position).to include([360, 177, "Bottom Member"])
       end
     end
 
@@ -138,117 +142,79 @@ describe Export::Pdf::Messages::LetterWithInvoice do
         let(:options) { {stamped: true} }
 
         it "renders only some texts positions" do
-          expect(text_with_position.count).to eq 58
-          expect(text_with_position).to match_array [
-            [71, 654, "Bottom Member"],
-            [71, 644, "Greatstreet 345"],
-            [71, 633, "3456 Greattown"],
-            [28, 290, "Empfangsschein"],
-            [28, 265, "Konto / Zahlbar an"],
-            [28, 254, "CH93 0076 2011 6238 5295 7"],
-            [28, 242, "Hans Gerber"],
-            [28, 230, "Eine Strasse 42"],
-            [28, 230, "Eine Strasse 42"],
-            [28, 219, "1234 Dorf"],
-            [28, 219, "1234 Dorf"],
-            [28, 187, "Zahlbar durch"],
-            [28, 175, "Bottom Member"],
-            [28, 164, "Greatstreet 345"],
-            [28, 152, "3456 Greattown"],
-            [28, 103, "Währung"],
-            [85, 103, "Betrag"],
-            [28, 92, "CHF"],
-            [85, 92, "10.00"],
-            [119, 53, "Annahmestelle"],
-            [360, 292, "Konto / Zahlbar an"],
-            [360, 280, "CH93 0076 2011 6238 5295 7"],
-            [360, 269, "Hans Gerber"],
-            [360, 225, "Referenznummer"],
-            [360, 214, "00 00834 96356 70000 00000 00019"],
-            [360, 193, "Zahlbar durch"],
-            [360, 181, "Bottom Member"],
-            [360, 170, "Greatstreet 345"],
-            [360, 158, "3456 Greattown"],
+          invoice_text = [
             [71, 654, "Top Leader"],
             [71, 644, "Funkystreet 42"],
             [71, 633, "4242 Greattown"],
             [28, 290, "Empfangsschein"],
             [28, 265, "Konto / Zahlbar an"],
-            [28, 254, "CH93 0076 2011 6238 5295 7"],
-            [28, 242, "Hans Gerber"],
-            [28, 187, "Zahlbar durch"],
-            [28, 175, "Top Leader"],
-            [28, 164, "Funkystreet 42"],
+            [28, 255, "CH93 0076 2011 6238 5295 7"],
+            [28, 246, "Hans Gerber"],
+            [28, 237, "Eine Strasse 42"],
+            [28, 228, "1234 Dorf"],
+            [28, 208, "Referenznummer"],
+            [28, 199, "00 00834 96356 70000 00000 00019"],
+            [28, 180, "Zahlbar durch"],
+            [28, 170, "Top Leader"],
+            [28, 161, "Funkystreet 42"],
             [28, 152, "4242 Greattown"],
             [28, 103, "Währung"],
             [85, 103, "Betrag"],
-            [28, 92, "CHF"],
-            [85, 92, "10.00"],
+            [28, 91, "CHF"],
+            [85, 91, "10.00"],
             [119, 53, "Annahmestelle"],
-            [360, 292, "Konto / Zahlbar an"],
-            [360, 280, "CH93 0076 2011 6238 5295 7"],
-            [360, 269, "Hans Gerber"],
-            [360, 257, "Eine Strasse 42"],
-            [360, 257, "Eine Strasse 42"],
-            [360, 246, "1234 Dorf"],
-            [360, 246, "1234 Dorf"],
-            [360, 225, "Referenznummer"],
-            [360, 214, "00 00834 96356 70000 00000 00019"],
-            [360, 193, "Zahlbar durch"],
-            [360, 181, "Top Leader"],
-            [360, 170, "Funkystreet 42"],
-            [360, 158, "4242 Greattown"]
+            [360, 290, "Konto / Zahlbar an"],
+            [360, 279, "CH93 0076 2011 6238 5295 7"],
+            [360, 267, "Hans Gerber"],
+            [360, 255, "Eine Strasse 42"],
+            [360, 244, "1234 Dorf"],
+            [360, 222, "Referenznummer"],
+            [360, 210, "00 00834 96356 70000 00000 00019"],
+            [360, 189, "Zahlbar durch"],
+            [360, 177, "Top Leader"],
+            [360, 165, "Funkystreet 42"],
+            [360, 154, "4242 Greattown"],
+            [71, 654, "Bottom Member"],
+            [71, 644, "Greatstreet 345"],
+            [71, 633, "3456 Greattown"],
+            [28, 290, "Empfangsschein"],
+            [28, 265, "Konto / Zahlbar an"],
+            [28, 255, "CH93 0076 2011 6238 5295 7"],
+            [28, 246, "Hans Gerber"],
+            [28, 237, "Eine Strasse 42"],
+            [28, 228, "1234 Dorf"],
+            [28, 208, "Referenznummer"],
+            [28, 199, "00 00834 96356 70000 00000 00019"],
+            [28, 180, "Zahlbar durch"],
+            [28, 170, "Bottom Member"],
+            [28, 161, "Greatstreet 345"],
+            [28, 152, "3456 Greattown"],
+            [28, 103, "Währung"],
+            [85, 103, "Betrag"],
+            [28, 91, "CHF"],
+            [85, 91, "10.00"],
+            [119, 53, "Annahmestelle"],
+            [360, 290, "Konto / Zahlbar an"],
+            [360, 279, "CH93 0076 2011 6238 5295 7"],
+            [360, 267, "Hans Gerber"],
+            [360, 255, "Eine Strasse 42"],
+            [360, 244, "1234 Dorf"],
+            [360, 222, "Referenznummer"],
+            [360, 210, "00 00834 96356 70000 00000 00019"],
+            [360, 189, "Zahlbar durch"],
+            [360, 177, "Bottom Member"],
+            [360, 165, "Greatstreet 345"],
+            [360, 154, "3456 Greattown"]
           ]
+
+          invoice_text.each_with_index do |text, i|
+            expect(text_with_position[i]).to eq(text)
+          end
         end
       end
 
       it "renders all texts at positions" do
-        expect(text_with_position.count).to eq 80
-        expect(text_with_position).to match_array [
-          [71, 654, "Bottom Member"],
-          [71, 644, "Greatstreet 345"],
-          [71, 633, "3456 Greattown"],
-          [71, 531, "Mitgliedsbeitrag"],
-          [71, 502, "Hallo"],
-          [71, 481, "Dein "],
-          [92, 481, "Mitgliedsbeitrag"],
-          [161, 481, " ist fällig! "],
-          [71, 460, "Bis bald"],
-          [28, 290, "Empfangsschein"],
-          [28, 265, "Konto / Zahlbar an"],
-          [28, 254, "CH93 0076 2011 6238 5295 7"],
-          [28, 242, "Hans Gerber"],
-          [28, 230, "Eine Strasse 42"],
-          [28, 230, "Eine Strasse 42"],
-          [28, 219, "1234 Dorf"],
-          [28, 219, "1234 Dorf"],
-          [28, 187, "Zahlbar durch"],
-          [28, 175, "Bottom Member"],
-          [28, 164, "Greatstreet 345"],
-          [28, 152, "3456 Greattown"],
-          [28, 103, "Währung"],
-          [85, 103, "Betrag"],
-          [28, 92, "CHF"],
-          [85, 92, "10.00"],
-          [119, 53, "Annahmestelle"],
-          [204, 290, "Zahlteil"],
-          [204, 103, "Währung"],
-          [261, 103, "Betrag"],
-          [204, 92, "CHF"],
-          [261, 92, "10.00"],
-          [360, 292, "Konto / Zahlbar an"],
-          [360, 280, "CH93 0076 2011 6238 5295 7"],
-          [360, 269, "Hans Gerber"],
-          [360, 257, "Eine Strasse 42"],
-          [360, 257, "Eine Strasse 42"],
-          [360, 246, "1234 Dorf"],
-          [360, 246, "1234 Dorf"],
-          [360, 225, "Referenznummer"],
-          [360, 214, "00 00834 96356 70000 00000 00019"],
-          [360, 193, "Zahlbar durch"],
-          [360, 181, "Bottom Member"],
-          [360, 170, "Greatstreet 345"],
-          [360, 158, "3456 Greattown"],
+        invoice_text = [
           [71, 654, "Top Leader"],
           [71, 644, "Funkystreet 42"],
           [71, 633, "4242 Greattown"],
@@ -260,39 +226,85 @@ describe Export::Pdf::Messages::LetterWithInvoice do
           [71, 460, "Bis bald"],
           [28, 290, "Empfangsschein"],
           [28, 265, "Konto / Zahlbar an"],
-          [28, 254, "CH93 0076 2011 6238 5295 7"],
-          [28, 242, "Hans Gerber"],
-          [28, 187, "Zahlbar durch"],
-          [28, 175, "Top Leader"],
-          [28, 164, "Funkystreet 42"],
+          [28, 255, "CH93 0076 2011 6238 5295 7"],
+          [28, 246, "Hans Gerber"],
+          [28, 237, "Eine Strasse 42"],
+          [28, 228, "1234 Dorf"],
+          [28, 208, "Referenznummer"],
+          [28, 199, "00 00834 96356 70000 00000 00019"],
+          [28, 180, "Zahlbar durch"],
+          [28, 170, "Top Leader"],
+          [28, 161, "Funkystreet 42"],
           [28, 152, "4242 Greattown"],
           [28, 103, "Währung"],
           [85, 103, "Betrag"],
-          [28, 92, "CHF"],
-          [85, 92, "10.00"],
+          [28, 91, "CHF"],
+          [85, 91, "10.00"],
           [119, 53, "Annahmestelle"],
           [204, 290, "Zahlteil"],
-          [204, 103, "Währung"],
-          [261, 103, "Betrag"],
-          [204, 92, "CHF"],
-          [261, 92, "10.00"],
-          [360, 292, "Konto / Zahlbar an"],
-          [360, 280, "CH93 0076 2011 6238 5295 7"],
-          [360, 269, "Hans Gerber"],
-          [360, 225, "Referenznummer"],
-          [360, 214, "00 00834 96356 70000 00000 00019"],
-          [360, 193, "Zahlbar durch"],
-          [360, 181, "Top Leader"],
-          [360, 170, "Funkystreet 42"],
-          [360, 158, "4242 Greattown"]
+          [204, 102, "Währung"],
+          [261, 102, "Betrag"],
+          [204, 90, "CHF"],
+          [261, 90, "10.00"],
+          [360, 290, "Konto / Zahlbar an"],
+          [360, 279, "CH93 0076 2011 6238 5295 7"],
+          [360, 267, "Hans Gerber"],
+          [360, 255, "Eine Strasse 42"],
+          [360, 244, "1234 Dorf"],
+          [360, 222, "Referenznummer"],
+          [360, 210, "00 00834 96356 70000 00000 00019"],
+          [360, 189, "Zahlbar durch"],
+          [360, 177, "Top Leader"],
+          [360, 165, "Funkystreet 42"],
+          [360, 154, "4242 Greattown"],
+          [71, 654, "Bottom Member"],
+          [71, 644, "Greatstreet 345"],
+          [71, 633, "3456 Greattown"],
+          [71, 531, "Mitgliedsbeitrag"],
+          [71, 502, "Hallo"],
+          [71, 481, "Dein "],
+          [92, 481, "Mitgliedsbeitrag"],
+          [161, 481, " ist fällig! "],
+          [71, 460, "Bis bald"],
+          [28, 290, "Empfangsschein"],
+          [28, 265, "Konto / Zahlbar an"],
+          [28, 255, "CH93 0076 2011 6238 5295 7"],
+          [28, 246, "Hans Gerber"],
+          [28, 237, "Eine Strasse 42"],
+          [28, 228, "1234 Dorf"],
+          [28, 208, "Referenznummer"],
+          [28, 199, "00 00834 96356 70000 00000 00019"],
+          [28, 180, "Zahlbar durch"],
+          [28, 170, "Bottom Member"],
+          [28, 161, "Greatstreet 345"],
+          [28, 152, "3456 Greattown"],
+          [28, 103, "Währung"],
+          [85, 103, "Betrag"],
+          [28, 91, "CHF"],
+          [85, 91, "10.00"],
+          [119, 53, "Annahmestelle"],
+          [204, 290, "Zahlteil"],
+          [204, 102, "Währung"],
+          [261, 102, "Betrag"],
+          [204, 90, "CHF"],
+          [261, 90, "10.00"],
+          [360, 290, "Konto / Zahlbar an"],
+          [360, 279, "CH93 0076 2011 6238 5295 7"],
+          [360, 267, "Hans Gerber"],
+          [360, 255, "Eine Strasse 42"],
+          [360, 244, "1234 Dorf"],
+          [360, 222, "Referenznummer"],
+          [360, 210, "00 00834 96356 70000 00000 00019"],
+          [360, 189, "Zahlbar durch"],
+          [360, 177, "Bottom Member"],
+          [360, 165, "Greatstreet 345"],
+          [360, 154, "3456 Greattown"]
         ]
-      end
-    end
-  end
 
-  def text_with_position
-    subject.positions.each_with_index.collect do |p, i|
-      p.collect(&:round) + [subject.show_text[i]]
+        invoice_text.each_with_index do |text, i|
+          expect(text_with_position[i]).to eq(text)
+        end
+      end
     end
   end
 end